// client/src/app/edit-article/types/edit-article-state.interface.ts
import {BackendErrorsInterface} from '../../shared/types/backend-errors.interface'
import {ArticleInterface} from '../../shared/types/article.interface'

export interface EditArticleStateInterface {
  isLoading: boolean
  article: ArticleInterface | null
  isSubmitting: boolean
  validationErrors: BackendErrorsInterface | null
}